Switched priority types from strings to integers
Instead of tracking task priority with words like “high”, “low”, etc,
this commit changes task priorities to be tracked with positive
integers. Lower integers are higher priority. Why? Simplicity and
performance. It is much easier for a database to sort numbers than to
convert words to numbers and then sort. It also simplifies the
“vocabulary”, meaning the specific word used is not relevant - only the
magnitude of number.