Get or create task-def on task create request
For now, I believe it is acceptable for the task-service to
automatically create a task-def based off of a task creation request if
a task-def of the specified name doesn’t already exist.
This also helps with ensuring the core-service tests succeed in their
current form, as they do not create a task-def prior to creating a task.