Added endpoint for task failure; updated queue tests
I prefer the design pattern where a request hits an endpoint for a
specific task and the server does the updating, rather than the client
doing the updating and hitting a general purpose “update” endpoint
where the client can update any value. I followed this pattern in
532d1f8fc06f3ddca22b399dae9d058b0f933784 and continue it here by adding
in an endpoint for “failing” a task. I also updated the queue tests as
they used to hit the general purpose “update” endpoint - they now hit
the specific complete/fail/release/etc. endpoints.